Move language.rs out of renderer (into stub)
